execute bundle config
execute bundle config build.nokogiri --use-system-libraries, before
bundle install.

When upgrading from nokogiri 1.6.6.2 to 1.6.6.3, we were failing with errors like
this:

```
Gem::Ext::BuildError: ERROR: Failed to build gem native extension.

    current directory:
    /pipeline/cache/bundle-install/ruby/2.2.0/gems/nokogiri-1.6.6.3/ext/nokogiri
    /usr/local/bin/ruby -r ./siteconf20151117-17-kx1dnr.rb extconf.rb
    checking if the C compiler accepts ... yes
    Building nokogiri using packaged libraries.
    checking for gzdopen() in -lz... yes
    checking for iconv... yes
...

Team Nokogiri will keep on doing their best to provide security
updates in a timely manner, but if this is a concern for you and want
to use the system library instead; abort this installation process and
reinstall nokogiri as follows:

    gem install nokogiri -- --use-system-libraries
        [--with-xml2-config=/path/to/xml2-config]
        [--with-xslt-config=/path/to/xslt-config]
```